restructure file sys
