+++ app/(root)/articles/[id]/page.tsx
@@ -13,7 +13,7 @@ import { notFound } from 'next/navigation';
 import markdownit from 'markdown-it';
 import { Timer } from 'lucide-react';
 import SaveButton from '@/components/SaveButton';
-import { LikeButton } from '@/components/LikeButton';
+import { LikeButton } from '@/components/Buttons/LikeButton';
 import { Metadata, ResolvingMetadata } from 'next';
 
 type Props = {
