+++ package.json
@@ -50,6 +50,7 @@
     "langchain": "^0.3.19",
     "lucide-react": "^0.479.0",
     "markdown-it": "^14.1.0",
+    "motion": "^12.6.2",
     "next": "^15.2.3",
     "next-auth": "^5.0.0-beta.25",
     "next-sanity": "^9.8.57",
